Addresses Issue 74. Deleted the extra script file that wasn't being used anymore. Renamed the main extraction file to avoid further confusion. Added comments to depict what is going on. Added further error handling in case some images or slides are not being extracted properly.
